requirements.txt
streamlit
python-dotenv
wikipedia
PyPDF2
python-docx
python-pptx
langchain
sentence-transformers
faiss-cpu
transformers
huggingface-hub
google-generativeai
openai